requirements.txt
h5py==2.10.0
joblib==0.15.1
numpy==1.18.1
pandas==0.25.3
rgf-python==3.8.0
scikit-learn==0.22.1
scipy==1.4.1
tqdm==4.46.1

